Created class  InputConsole
Moved appropriate code to the new class InputConsole (which  extends
JTextPane)
